// src/types/product.ts
export type Product = {
  product_id: number;
  product_in_menu_id: number;
  product_name: string;
  product_name_en: string;
  bean: number;
  pic_url: string;
  price: number;
  base_price: number;
  product_type_id: number;
  product_type_name: string;
  category_id: number;
  category_name: string;
  supplier_id: number;
  supplier_name: string;
  product_code: string;
  // "attributes": {},
  // "child_products": [],
  // "extras": [],
  // "product_combination_models": []
};